+++ Thunderstore/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+### v1.0.1
+
+- Client have a random color when joining a game with scrap already in ship
+- Add a big jar of pickles (two handed)
+
 ### v1.0.0
 
 - Release version